Lots of changes
But it's all vibe coded, and I hate that kinda... also it's crappy code, but made it work so there's that. But at what point does vibing just become a proxy to writing stuff? I mean if i tell the AI to type A and it does that, is that vibe coding?